Removed dynamic member lookup in favor of named getters.
gzip'd versions of text resources are now generated and accessible, for
RELEASE builds specifically (we don't support gzip-ing on the fly for
DEBUG builds, as gzip compression is considered a release build
optimization. While the gzip'd versions of the static resources are
directly available in DEBUG builds, they are not included in the
dynamic resource lookup).